requirements.txt
pytorch >= 1.1.0
numpy >= 1.15.4 #For data preparation process
scikit-learn >= 0.20.2 #For AUC-PR
torch-cluster==1.4.3  
torch-geometric==1.3.0    
torch-scatter==1.3.1
torch-sparse==0.4.0  
networkx==2.3  
